convert all news articles state data from context to redux store
