+++ client/src/context/GlobalCryptoContext.tsx
@@ -1,4 +1,4 @@
-import { useContext, createContext, useState, useEffect } from "react";
+import { useContext, createContext, useState } from "react";
 import { PopupBanner } from "../components/PopupBanner";
 import { LoadingUx } from "../components/LoadingUx";
 import {
@@ -9,7 +9,6 @@ import {
 } from "./context.models";
 import { useUserAuth } from "../hooks/useUserAuth";
 import { useFetchPosts } from "../hooks/useFetchPosts";
-import { useFetchNews } from "../hooks/useFetchNews";
 
 // Create context, and export custom hook that can extract our context values in different components.
 const CryptoContext = createContext<ContextInterface | null>(null);
@@ -28,7 +27,6 @@ export const GlobalCryptoProvider: ContextProviderComponent = ({
   // Utilizes our custom hooks that fetch our application data
   const [user, setUser, token, setToken] = useUserAuth();
   const [cryptos, setCryptos, serverOffline] = useFetchPosts();
-  const newsArticles = useFetchNews();
 
   const [pageLoading, setPageLoading] = useState<boolean>(false);
   const [showPopupBanner, setShowPopupBanner] = useState<BannerMessage>({
@@ -56,7 +54,6 @@ export const GlobalCryptoProvider: ContextProviderComponent = ({
         setToken,
         togglePageLoading,
         handleBannerMessage,
-        newsArticles,
       }}
     >
       {/* Will populate all nested children components */}
