+++ client/src/features/store.ts
@@ -1,7 +1,8 @@
 import { configureStore } from "@reduxjs/toolkit";
+import { newsArticlesReducer } from "./newsArticles/newsArticlesSlice";
 
 export const store = configureStore({
-  reducer: {},
+  reducer: { newsArticles: newsArticlesReducer },
 });
 
 // Needed for typing our state and dispatch functions anytime they're used in out application.
