+++ client/src/index.tsx
@@ -3,14 +3,16 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import { RouteSwitch } from "./RouteSwitch";
 import { GlobalCryptoProvider } from "./context/GlobalCryptoContext";
+import { Provider } from "react-redux";
+import { store } from "./features/store";
 
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
 );
 root.render(
-  <React.StrictMode>
+  <Provider store={store}>
     <GlobalCryptoProvider>
       <RouteSwitch />
     </GlobalCryptoProvider>
-  </React.StrictMode>
+  </Provider>
 );
