+++ readme.md
@@ -56,3 +56,5 @@ API server would need the following environments and variables provided to run s
 - Add fourth statistic on price history on crypto card.
 
 > Bugs
+
+- Fix dark theme not filling up entire page on News route, leaving a partial white background.
